Store container in a static variable on CacheIncludeExtension
The container was being stored in an instance variable, which is a cleaner
implementation than what I've changed it to, however due to the way
SilverStripe extensions work, the container instance variable ends up
being part of every class the CacheIncludeExtension is applied to. When
the extension is applied to DataObject, it becomes impossible to use the
`serialize()` method on any DataObject.

The use of an instance variable to store the container is only used for
convenience at this point, so this tweak doesn't add any limitations yet.