Loosen version constraint on doctrine/cache
I'm working through an issue with CacheInclude's cache not clearing on
save on a production site. There've been quite a few bug fixes in
releases to doctrine/cache over the last year, and while I can't see a
fix that specifically address the issue I'm seeing, it's possible the
underlying cause of my issue has been fixed in somewhere in there.

Haven't been able to replicate the issue occurring on the production
site in production or locally, but it's occurred twice in the last day
by itself. `doctrine/cache` >1.3 changes the cache file/folder names
also, so it's not possible to test if updating this will actually fix
my issue, since it effectively uses a different cache.

This is slightly voodoo, but doctrine/cache uses semver, so it shouldn't
be a problem loosening this constraint.